Updates - webpack configured
Using cookie-session instead of express-session for lightweight cookie
sessions stored on the client since no sensitive information is used.

Manually cleared node_modules and reinstalled modules to fix some
errors.

Set the css-loader url option to false so that the images in the css
will be ignored and not bundle with webpack since they're not modified.